renamed files and exracted TwitterAPIClient to new file
